Move LeftPanel axios call back into action creator so that endpoint only gets requested once
