+++ react-redux-app/src/components/LeftPanel/LeftPanel.js
@@ -1,5 +1,4 @@
-import React, { useState, useEffect } from "react";
-import axios from "axios";
+import React from "react";
 import styled from "styled-components";
 import { connect } from "react-redux";
 
@@ -17,86 +16,14 @@ const NameScreen = styled(Screen)`
 `;
 
 const LeftPanel = props => {
-    const [dexEntries, setDexEntries] = useState([]);
-    const [dexNum, setDexNum] = useState(null);
-
-    // get dexEntries and dexNum
-    useEffect(() => {
-        let dexNumArray = [];
-        let natDexNum = null;
-        let entries = [];
-
-        if (props.pokemon !== {} && props.pokemon.species) {
-            axios
-                .get(props.pokemon.species.url)
-                .then(res => {
-                    // console.log("leftPanel comp state species url res: ", res);
-
-                    // array of dex objects for different pokedexes
-                    dexNumArray = [...res.data["pokedex_numbers"]];
-
-                    // find the national dex object
-                    natDexNum = dexNumArray.length
-                        ? dexNumArray.find(
-                              obj => obj.pokedex.name === "national"
-                          )
-                        : "N/A";
-
-                    // get just the national dex number
-                    if (natDexNum !== "N/A") {
-                        natDexNum = natDexNum["entry_number"];
-                    }
-
-                    entries = [...res.data["flavor_text_entries"]];
-
-                    // only keep the english entries
-                    entries = entries.filter(
-                        entry => entry.language.name === "en"
-                    );
-
-                    // make all whitespace consistent by making them all spaces
-                    entries.forEach(entry => {
-                        entry["flavor_text"] = entry["flavor_text"].replace(
-                            /\s/gm,
-                            " "
-                        );
-                    });
-
-                    // find and get rid of duplicate entries
-                    entries = entries.reduce((acc, current) => {
-                        const x = acc.find(
-                            item =>
-                                item["flavor_text"] === current["flavor_text"]
-                        );
-                        if (!x) {
-                            return acc.concat([current]);
-                        } else {
-                            return acc;
-                        }
-                    }, []);
-
-                    setDexEntries(entries);
-                    setDexNum(natDexNum);
-                })
-                .catch(err => {
-                    console.log(
-                        "Error getting species url in LeftPanel: ",
-                        err
-                    );
-                });
-        }
-    }, [props.pokemon]);
-
-    // useEffect(() => {
-    //     console.log("LeftPanel: isFetching: ", props.isFetching);
-    //     console.log("pokemon: ", props.pokemon);
-    // }, [props.isFetching]);
-
     return (
         <Panel>
             <NameScreen>
                 {props.pokemon !== {} && !props.isFetching && (
-                    <NameScreenText name={props.pokemon.name} dexNum={dexNum} />
+                    <NameScreenText
+                        name={props.pokemon.name}
+                        dexNum={props.dexNum}
+                    />
                 )}
             </NameScreen>
             <MainSprite
@@ -104,7 +31,10 @@ const LeftPanel = props => {
                 name={props.pokemon.name}
                 isFetching={props.isFetching}
             />
-            <DexEntries dexEntries={dexEntries} isFetching={props.isFetching} />
+            <DexEntries
+                dexEntries={props.dexEntries}
+                isFetching={props.isFetching}
+            />
         </Panel>
     );
 };
@@ -112,6 +42,8 @@ const LeftPanel = props => {
 const mapStateToProps = state => {
     return {
         pokemon: state.pokemon,
+        dexEntries: state.dexEntries,
+        dexNum: state.dexNum,
         isFetching: state.isFetching
     };
 };
